fix2.java
return(double)..

+++ src/main/java/com/thealgorithms/maths/Average.java
@@ -45,6 +45,6 @@ public final class Average {
         for (int number : numbers) {
             sum += number;
         }
-        return (int) (sum / numbers.length);
+        return (double) (sum / numbers.length);
     }
 }
